initial
There is a Dessert Shoppe which sells cookies by the dozen, ice cream, and sundaes (ice cream with a topping).
Your code will be used for the checkout system.
To do this, you will implement an inheritance hierarchy of classes derived from a DessertItem superclass.
The Cookie, and IceCream classes will be derived from the DessertItem class.
The Sundae class will be derived from the IceCream class.
The DessertItem Class
The DessertItem class is a superclass from which specific types of DessertItems can be derived. It contains only
one data member, a name (char *). It also defines a number of methods. All of the DessertItem class methods
except the getCost() method are defined in a generic way.
The getCost() method in DessertItem class should only return 0 because the method of determining the costs
varies based on the type of item.
Tax amounts should be rounded to the nearest Paisa. For example, calculating the tax on a food item with a cost
of 199 paisa with a tax rate of 2.0% should be 4 paisa.
The Derived Classes
All of the classes which are derived from the DessertItem class must define a constructor.
The Cookie class should be derived from the DessertItem class. A Cookie item has a number and a price per
dozen which are used to determine its cost. For example, 4 cookies @ 399 paisa /dz. = 133 paisa. The cost
should be rounded to the nearest paisa.
The IceCream class should be derived from the DessertItem class. An IceCream item simply has a cost.
The Sundae class should be derived from the IceCream class. The cost of a Sundae is the cost of the IceCream
plus the cost of the topping.